Change newPrice type from int to decimal for UpdatePrice
Updated the UpdatePrice method in both ProductController and ProductService to use decimal instead of int for the newPrice parameter, enabling support for fractional price values.

+++ FirstWebAPI/Controllers/ProductController.cs
@@ -36,7 +36,7 @@ namespace FirstWebApp.Controllers
         }
 
         [HttpPut("UpdatePrice")]
-        public IActionResult UpdatePrice(int productId, int newPrice)
+        public IActionResult UpdatePrice(int productId, decimal newPrice)
         {
             bool updated = productService.UpdatePrice(productId, newPrice);
